Added lib/types.
