+++ queries/analytics/pageview/getPageviewMetrics.ts
@@ -3,7 +3,7 @@ import clickhouse from 'lib/clickhouse';
 import { runQuery, CLICKHOUSE, PRISMA } from 'lib/db';
 import cache from 'lib/cache';
 import { Prisma } from '@prisma/client';
-import { UmamiApi } from 'lib/constants';
+import { EventType } from 'lib/types';
 
 export async function getPageviewMetrics(
   ...args: [
@@ -43,7 +43,7 @@ async function relationalQuery(
       ${joinSession}
     where website_id='${websiteId}'
       and website_event.created_at between $1 and $2
-      and event_type = ${UmamiApi.EventType.Pageview}
+      and event_type = ${EventType.Pageview}
       ${filterQuery}
     group by 1
     order by 2 desc`,
@@ -71,7 +71,7 @@ async function clickhouseQuery(
     from event
     where website_id = $1
       and rev_id = $2
-      and event_type = ${UmamiApi.EventType.Pageview}
+      and event_type = ${EventType.Pageview}
       ${column !== 'event_name' ? `and event_name = ''` : `and event_name != ''`}
       and ${getBetweenDates('created_at', startDate, endDate)}
       ${filterQuery}
